ensure prediction only runs if update runs, change type of previous timestamp to long long, and change division of timestamp

+++ src/ukf.h
@@ -75,7 +75,7 @@ public:
   double NIS_laser_;
 
   // Initialize time
-  float previous_timestamp_ = 0.0;
+  long long previous_timestamp_ = 0;
 
   /**
    * Constructor
